Added new properties to OldOrderModel, new subroutines to OldOrderPresenter, OldOrderInterface.
New subs in OldOrderInterface will probably be gone in next update.

Added file stream object to RecordConverter.vb, will be used to represent and work with OldOrder files in the near future. Added comments outlining the general execution flow of this process.

Added text fields and buttons to form, so as to solidify inputs and outputs the code is being designed around.