Added functionality to get multiple record types from OldOrder files - Customer, Item, Total, and Notes.
Added variables to signal to views what kind of record is being worked with, properties so that other view types know if they can begin working on their records.

Corrected variable naming syntax - all private variables now have p_ as a prefix (for private). Before it was p_ for private and presenter, and m_ for model variable, this didn't make sense in the long run.