Primary changes:
Updated to only write to one output file.

Data integrity, ID creation, and duplicate checking will be done in SQL statements.

Updated UI to show import file, error log, and failed read file locations.

Removed unnecessary presenter and model references from the main body code.

Updated manner in which output file is written to, data is now written to list of structure, so go/no go decisions on whether or not to write data to import file can be made if the old order is found to be bad during the read.

Cleaned up property, private variable names so as to make more sense.

Added fields to provide more information regarding the order, order item, and time of order.